install meson from scratch on ClusterFuzz only

+++ tools/oss-fuzz.sh
@@ -32,8 +32,10 @@ if [ -z "$FUZZING_ENGINE" ]; then
     fuzzflag="llvm-fuzz=true"
 fi
 
+if [[ -n "$LIB_FUZZING_ENGINE" ]]; then
 # link_language: 'cpp' (introduced in https://github.com/mesonbuild/meson/pull/4978) hasn't been released yet
-pip3 uninstall --yes meson && git clone https://github.com/mesonbuild/meson && cd meson && python3 setup.py install && cd .. && rm -rf ./meson
+    pip3 uninstall --yes meson && git clone https://github.com/mesonbuild/meson && cd meson && python3 setup.py install && cd .. && rm -rf ./meson
+fi
 
 # we should probably pass cap=false and mount=false only on "i386"
 meson $build -Dcap=false -Dmount=false -D$fuzzflag -Db_lundef=false
